daemon/c_seccomp: Moved finit_module emulation code to seperate file
This moves the emulation code for the finit_module() syscall into the
separate file 'c_seccomp/init_module.c'. We did not change the
emulation code itself. Just moved c_seccomp_emulate_finit_module() to
the new file and generated necessary interface code.
